Add class variable to return Scheduler exit code
Add a class variable `_return_scheduler_exitcode` to the Calcjob class.
If this variable is set to True, the parse method will return any
non-zero exit codes found by the parse_output method of the Scheduler.
Since this is set to False by default, this addition has no influence on
how the class operates, but in case a plugin developer would like to have
this feature enabled, he/she can set the `_return_scheduler_exitcode`
to `True` for the calculation job class (e.g. `PwCalculation`).