Homogenize the attributes/extras methods
Make sure the code for the attributes and extras methods are
identical, as a first step towards refactoring the code to use a
mixin class for these methods. These changes should have no
influence on how the methods function.

Add exception chaining for exceptions raised directly during the
handling of another exception. There is only a minor difference in
the output, but it should make it clear that this exception was
raised purposefully.