Add mixin classes for attributes/extras methods
Move the attributes and extras methods to two mixin classes called
`AttributesBackendEntity` and `ExtrasBackendEntity`. The mixin
classes rely on the `is_stored` and `_flush_if_stored` methods, so
these are added as abstract methods. They are "mixed in" at the
BackendNode` level, where the abstract methods of the attributes and
extras are removed.

Move the `_flush_if_stored` method to the `BackendEntity` class,
which is added leftmost to the `BackendNode` parent classes. This
method can be used by all backend entities.

Move the `validates_attribute_extra_key` method from the frontend
`Node` class to the backend mixin classes `AttributesBackendEntity`
and `ExtrasBackendEntity`. This way the key/value
validation/cleaning is both done at the backend level, which is
more consistent. Moreover, this means other frontend classes won't
have to add this call to `validates_attribute_extra_key` to their
methods, when they want to use the attributes/extras methods.