Add extras to Group
Add extras column to the database models for groups, with an empty
dictionary as a default. Add the `ExtrasBackendEntity` mixin class
to the `BackendGroup` to make sure the group backend classes for
django and sqlalchemy have the extras methods. Add the
corresponding extras methods to the frontend `Group` class.

For SqlAlchemy: instead of using the constructor to set the default
values of the attributes and extras for `DbNode` and `DbGroup`, use
the Column declaration to do so. Generate the required migration
files and make adjustments where necessary.

Also fix some typos in the utils.py modules and add exception
chaining for exceptions raised straight from the handling of a
previous exception.